Add logic for cleaning all PE
