+++ .github/actions/cleanup-pe/cleanPE.go
@@ -6,29 +6,21 @@ import (
 	"log"
 	"os"
 
-	"github.com/Azure/azure-sdk-for-go/profiles/latest/network/mgmt/network"
-	"github.com/Azure/go-autorest/autorest/azure/auth"
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/session"
-	"github.com/aws/aws-sdk-go/service/ec2"
-	"google.golang.org/api/compute/v1"
-
 	"github.com/mongodb/mongodb-atlas-kubernetes/test/e2e/actions/cloud"
-	"github.com/mongodb/mongodb-atlas-kubernetes/test/e2e/api/gcp"
 	"github.com/mongodb/mongodb-atlas-kubernetes/test/e2e/config"
 	"github.com/mongodb/mongodb-atlas-kubernetes/test/e2e/utils"
 )
 
 func main() {
-	if os.Getenv("CLEAN_PE") != "true" {
-		log.Println("CLEAN_PE is not true. skipping cleanup PE")
-		return
+	cleanOnlyTaggedPE := false
+	if os.Getenv("CLEAN_PE") == "true" {
+		cleanOnlyTaggedPE = true
 	}
 	err := SetGCPCredentials()
 	if err != nil {
 		log.Fatal(err)
 	}
-	err = CleanAllPE()
+	err = CleanAllPE(cleanOnlyTaggedPE)
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -46,154 +38,52 @@ func SetGCPCredentials() error {
 	return nil
 }
 
-func CleanAllPE() error {
+func CleanAllPE(onlyTagged bool) error {
 	ctx := context.Background()
 	groupNameAzure := cloud.ResourceGroup
-	subscriptionID := os.Getenv("AZURE_SUBSCRIPTION_ID")
-	err := cleanAllAzurePE(ctx, config.TagForTestKey, config.TagForTestValue, groupNameAzure, subscriptionID)
-	if err != nil {
-		return fmt.Errorf("error while cleaning all azure pe: %v", err)
-	}
-
 	awsRegions := []string{
 		config.AWSRegionEU,
 		config.AWSRegionUS,
 	}
-	for _, awsRegion := range awsRegions {
-		errClean := cleanAllAWSPE(awsRegion, config.TagForTestKey, config.TagForTestValue)
-		if errClean != nil {
-			return fmt.Errorf("error cleaning all aws PE. region %s. error: %v", awsRegion, errClean)
-		}
-	}
-
 	gcpRegion := config.GCPRegion
-	err = cleanAllGCPPE(ctx, cloud.GoogleProjectID, cloud.GoogleVPC,
-		gcpRegion, config.TagForTestKey, config.TagForTestValue)
-	if err != nil {
-		return fmt.Errorf("error while cleaning all gcp pe: %v", err)
-	}
-	return nil
-}
+	subscriptionID := os.Getenv("AZURE_SUBSCRIPTION_ID")
 
-func cleanAllAWSPE(region, tagName, tagValue string) error {
-	awsSession, err := session.NewSession(&aws.Config{
-		Region: aws.String(region)},
-	)
-	if err != nil {
-		return fmt.Errorf("error creating awsSession: %v", err)
-	}
-	svc := ec2.New(awsSession)
-	endpoints, err := svc.DescribeVpcEndpoints(&ec2.DescribeVpcEndpointsInput{
-		Filters: []*ec2.Filter{
-			{
-				Name: aws.String("tag:Name"),
-				Values: []*string{
-					aws.String(tagName),
-				},
-			},
-			{
-				Name: aws.String("tag:Value"),
-				Values: []*string{
-					aws.String(tagValue),
-				},
-			},
-		},
-	})
-	if err != nil {
-		return fmt.Errorf("error fething all vpcEP with tag %s: %v", tagName, err)
-	}
-	var endpointIDs []*string
-	for _, endpoint := range endpoints.VpcEndpoints {
-		endpointIDs = append(endpointIDs, endpoint.VpcEndpointId)
-	}
-	if len(endpointIDs) > 0 {
-		input := &ec2.DeleteVpcEndpointsInput{
-			VpcEndpointIds: endpointIDs,
-		}
-		_, err = svc.DeleteVpcEndpoints(input)
+	if onlyTagged {
+		err := cleanAllTaggedAzurePE(ctx, config.TagForTestKey, config.TagForTestValue, groupNameAzure, subscriptionID)
 		if err != nil {
-			return fmt.Errorf("error deleting vpcEP: %v", err)
+			return fmt.Errorf("error while cleaning all azure pe: %v", err)
 		}
-	}
-	log.Printf("deleted %d AWS PEs in region %s", len(endpointIDs), region)
-	return nil
-}
 
-func cleanAllAzurePE(ctx context.Context, tagName, tagValue, resourceGroupName, azureSubscriptionID string) error {
-	authorizer, err := auth.NewAuthorizerFromEnvironment()
-	if err != nil {
-		return fmt.Errorf("error creating authorizer: %v", err)
-	}
-	peClient := network.NewPrivateEndpointsClient(azureSubscriptionID)
-	peClient.Authorizer = authorizer
-
-	peList, err := peClient.List(ctx, resourceGroupName)
-	if err != nil {
-		return fmt.Errorf("error fething all PE: %v", err)
-	}
-	var endpointNames []string
-	for _, endpoint := range peList.Values() {
-		tags := endpoint.Tags
-		if peTagValue, ok := tags[tagName]; ok {
-			if peTagValue != nil {
-				if *peTagValue == tagValue {
-					endpointNames = append(endpointNames, *endpoint.Name)
-				}
+		for _, awsRegion := range awsRegions {
+			errClean := cleanAllTaggedAWSPE(awsRegion, config.TagForTestKey, config.TagForTestValue)
+			if errClean != nil {
+				return fmt.Errorf("error cleaning all aws PE. region %s. error: %v", awsRegion, errClean)
 			}
 		}
-	}
 
-	for _, peName := range endpointNames {
-		_, errDelete := peClient.Delete(ctx, resourceGroupName, peName)
-		if errDelete != nil {
-			return errDelete
+		err = cleanAllTaggedGCPPE(ctx, cloud.GoogleProjectID, cloud.GoogleVPC,
+			gcpRegion, config.TagForTestKey, config.TagForTestValue)
+		if err != nil {
+			return fmt.Errorf("error while cleaning all gcp pe: %v", err)
 		}
-		log.Printf("successfully deleted Azure PE %s", peName)
-	}
-	return nil
-}
+	} else {
+		err := cleanAllAzurePE(ctx, groupNameAzure, subscriptionID)
+		if err != nil {
+			return fmt.Errorf("error while cleaning all azure pe: %v", err)
+		}
 
-func cleanAllGCPPE(ctx context.Context, projectID, vpc, region, tagName, tagValue string) error {
-	computeService, err := compute.NewService(ctx)
-	if err != nil {
-		return fmt.Errorf("error while creating new compute service: %v", err)
-	}
-
-	networkURL := gcp.FormNetworkURL(vpc, projectID)
-
-	forwardRules, err := computeService.ForwardingRules.List(projectID, region).Do()
-	if err != nil {
-		return fmt.Errorf("error while listing forwarding rules: %v", err)
-	}
-
-	for _, forwardRule := range forwardRules.Items {
-		forwardRuleLabels := forwardRule.Labels
-		if forwardRuleLabels[tagName] == tagValue && forwardRule.Network == networkURL {
-			_, err = computeService.ForwardingRules.Delete(projectID, region, forwardRule.Name).Do()
-			if err != nil {
-				return fmt.Errorf("error while deleting forwarding rule: %v", err)
-			}
-			ruleName := forwardRule.Name
-			log.Printf("successfully deleted GCP forward rule: %s", ruleName)
-			err = deleteGCPAddressByForwardRuleName(computeService, projectID, region, ruleName)
-			if err != nil {
-				return err
+		for _, awsRegion := range awsRegions {
+			errClean := cleanAllAWSPE(awsRegion)
+			if errClean != nil {
+				return fmt.Errorf("error cleaning all aws PE. region %s. error: %v", awsRegion, errClean)
 			}
 		}
-	}
 
-	return nil
-}
-
-func deleteGCPAddressByForwardRuleName(service *compute.Service, projectID, region, ruleName string) error {
-	addressName, err := cloud.FormAddressNameByRuleName(ruleName)
-	if err != nil {
-		return fmt.Errorf("unexpected forvard rule name pattern: %v", err)
+		err = cleanAllGCPPE(ctx, cloud.GoogleProjectID, cloud.GoogleVPC,
+			gcpRegion, cloud.GoogleSubnetName)
+		if err != nil {
+			return fmt.Errorf("error while cleaning all gcp pe: %v", err)
+		}
 	}
-	_, err = service.Addresses.Delete(projectID, region, addressName).Do()
-	if err != nil {
-		return fmt.Errorf("error while deleting address: %v", err)
-	}
-	log.Printf("successfully deleted GCP address: %s", addressName)
 	return nil
 }
