+++ .github/workflows/cleanup-all.yml
@@ -13,8 +13,8 @@ on:
         description: "Timelife of project in hours"
         default: "3"
         required: true
-      cleanPE:
-        description: "Clean PE(true) or not(false)?"
+      cleanTaggedPE:
+        description: "Clean only tagged PE(true) or all PE(false)?"
         default: "false"
         required: true
 
@@ -55,5 +55,5 @@ jobs:
           AZURE_CLIENT_SECRET: ${{ secrets.AZURE_CLIENT_SECRET }}
           AZURE_SUBSCRIPTION_ID: ${{ secrets.AZURE_SUBSCRIPTION_ID }}
           GCP_SA_CRED: ${{ secrets.GCP_SA_CRED }}
-          CLEAN_PE: ${{ github.event.inputs.cleanPE }}
+          CLEAN_PE: ${{ github.event.inputs.cleanTaggedPE }}
         run: go run .github/actions/cleanup-pe/cleanPE.go
